refactor: unify filter panel look across index, app, stats into shared filters.css
