+++ app.html
@@ -15,10 +15,11 @@
 <link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
 <link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
 <link rel="stylesheet" href="css/shared/site.css?v=4d518a51">
+<link rel="stylesheet" href="css/shared/filters.css?v=3b0c60ff">
 <link rel="stylesheet" href="css/shared/cards.css?v=68f59a72">
 <link rel="stylesheet" href="css/app/game-header.css?v=cec66a65">
 <link rel="stylesheet" href="css/app/configs.css?v=90d5bed5">
-<link rel="stylesheet" href="css/app/reports.css?v=7c356ffd">
+<link rel="stylesheet" href="css/app/reports.css?v=f4016583">
 <link rel="stylesheet" href="css/app/modals.css?v=2172da2e">
 <link rel="stylesheet" href="css/app/home.css?v=d8aa3fe5">
 </head>
@@ -57,6 +58,6 @@
 <script src="js/lib/topbar.js?v=50b75e70"></script>
 <script src="js/lib/toast.js?v=7bb23014"></script>
 <script src="js/lib/gh-gist.js?v=346e3ea7"></script>
-<script type="module" src="js/app/main.js?v=f55cb956"></script>
+<script type="module" src="js/app/main.js?v=96c048a9"></script>
 </body>
 </html>
