+++ css/index/index.css
@@ -314,60 +314,12 @@
 .pg-section-heading { margin: 6px 0 2px; }
 /* Unified Filters popover for the popular section */
 .pg-filter-wrap { position: relative; }
-.pg-filter-toggle-btn {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  padding: 6px 14px;
-  background: var(--s1);
-  border: 1.5px solid var(--border);
-  color: var(--muted);
-  font-size: 0.8rem;
-  font-weight: 700;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-  cursor: pointer;
-  border-radius: 999px;
-  white-space: nowrap;
-  transition: border-color 0.15s, color 0.15s, background 0.15s;
-}
-.pg-filter-toggle-btn:hover { border-color: var(--accent); color: var(--strong); }
-.pg-filter-toggle-btn.has-filters { color: var(--accent); border-color: var(--accent); background: rgba(74,159,208,0.08); }
-.pg-filter-badge {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  min-width: 18px;
-  height: 16px;
-  padding: 0 4px;
-  border-radius: 999px;
-  background: var(--accent);
-  color: #0a0c10;
-  font-size: 0.65rem;
-  font-weight: 800;
-}
-.pg-filter-panel {
-  display: none;
-  position: absolute;
-  top: calc(100% + 6px);
-  left: 0;
-  z-index: 200;
-  background: var(--s1);
-  border: 1px solid var(--border);
-  border-radius: 10px;
-  box-shadow: 0 8px 24px rgba(0,0,0,0.4);
-  padding: 14px 16px;
-  min-width: 200px;
-  flex-direction: column;
-  gap: 14px;
-}
+/* .pg-filter-toggle-btn, .pg-filter-badge -- base styles in css/shared/filters.css */
+/* .pg-filter-panel base -- in css/shared/filters.css */
+.pg-filter-panel { display: none; flex-direction: column; gap: 14px; }
 .pg-filter-panel.open { display: flex; }
 @media (min-width: 601px) {
-  .pg-filter-panel.open {
-    flex-direction: row;
-    gap: 24px;
-    min-width: 420px;
-  }
+  .pg-filter-panel.open { flex-direction: row; gap: 24px; }
 }
 .pg-filter-group {
   display: flex;
