+++ game-stats.html
@@ -15,7 +15,7 @@
 <link rel="stylesheet" href="css/shared/cards.css?v=68f59a72">
 <link rel="stylesheet" href="css/app/game-header.css?v=cec66a65">
 <link rel="stylesheet" href="css/app/configs.css?v=90d5bed5">
-<link rel="stylesheet" href="css/app/reports.css?v=7c356ffd">
+<link rel="stylesheet" href="css/app/reports.css?v=f4016583">
 <link rel="stylesheet" href="css/app/modals.css?v=2172da2e">
 <link rel="stylesheet" href="css/app/home.css?v=d8aa3fe5">
 <style>
