+++ gh-pages-manifest.txt
@@ -39,6 +39,7 @@ css/shared/base.css
 css/shared/topbar.css
 css/shared/site.css
 css/shared/cards.css
+css/shared/filters.css
 about.html
 options.html
 js/options/main.js
