+++ js/app/main.js
@@ -1,7 +1,7 @@
 // Entry point for the app page: bootstraps routing and search wiring.
 // (Replaces the inline bootstrap that lived at the top/bottom of app.js.)
-import { route } from './router.js?v=4f2ac3a6';
-import { wireSearch } from './components/search.js?v=f68f83d2';
+import { route } from './router.js?v=f6ca009b';
+import { wireSearch } from './components/search.js?v=69369686';
 
 window.addEventListener('hashchange', () => route());
 window.addEventListener('popstate', () => route());
