+++ privacy.html
@@ -12,7 +12,7 @@
 <link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
 <link rel="stylesheet" href="css/shared/site.css?v=4d518a51">
 <link rel="stylesheet" href="css/shared/cards.css?v=68f59a72">
-<link rel="stylesheet" href="css/index/index.css?v=07446e4c">
+<link rel="stylesheet" href="css/index/index.css?v=ca60e7d4">
 <style>
   .prose { max-width: 720px; margin: 2rem auto; line-height: 1.7; }
   .prose h1 { font-size: 1.6rem; margin-bottom: 0.5rem; }
